Test correct view is shown for single class method
Adding a test to caputure a bug that displays a blank view instead of
the methods view when the method only belongs to a single class.

When a method belongs to only one class then the expected behavior
should to automatically display the information for that class when the
user does not pass it in explicitly, however the current behavior
renders a blank view, and then states the method does not exist.