Convert symbol to string
Converting the symbol to a string as this variable is passed to a method
that compares it against other strings.

This is needed so that users that call a method without a class can
automatically see the information for that view when the method only
belongs to one class.

Fixes #2